LayoutTests/imported/w3c/web-platform-tests/editing/other/insertparagraph-in-child-of-html.tentative_designMode=off&white-space=normal-expected.txt



FAIL insertParagraph in empty <div style="white-space:normal"> after <body> should split the <div> assert_equals: The <div> should be split expected "<head><title>iframe</title></head><body><br></body><div style=\"white-space:normal\"><br></div><div style=\"white-space:normal\"><br></div>" but got "<head><title>iframe</title></head><body><br><div><br></div></body><div style=\"white-space:normal\"></div>"
PASS insertParagraph in <div style="white-space:normal"> (containing only a <br>) after <body> should split the <div>
PASS insertParagraph in <div style="white-space:normal"> (containing text) after <body> should not create another <div> element
FAIL insertParagraph in empty <div style="white-space:normal"> before <body> should split the <div> assert_equals: The <div> should be split expected "<head><title>iframe</title></head><div style=\"white-space:normal\"><br></div><div style=\"white-space:normal\"><br></div><body><br></body>" but got "<head><title>iframe</title></head><div style=\"white-space:normal\"></div><body><br><div><br></div></body>"
PASS insertParagraph in <div style="white-space:normal"> (containing only a <br>) before <body> should split the <div>
PASS insertParagraph in <div style="white-space:normal"> (containing text) before <body> should split the <div>
FAIL insertParagraph in empty <div style="white-space:normal"> before <head> should split the <div> assert_equals: The <div> should be split expected "<div style=\"white-space:normal\"><br></div><div style=\"white-space:normal\"><br></div><head><title>iframe</title></head><body><br></body>" but got "<div style=\"white-space:normal\"></div><head><title>iframe</title></head><body><br><div><br></div></body>"
PASS insertParagraph in <div style="white-space:normal"> (containing only a <br>) before <head> should split the <div>
PASS insertParagraph in <div style="white-space:normal"> (containing text) before <head> should split the <div>
